API: Split description messages into summary + additional text
See MediaWiki core patch I778bab2b

This change is intentionally written to not include the deprecated
message so new extensions using this an an example will be less likely
to include it.

This is a resubmission of I4b2b34ea which was force-merged without the
necessary core change, and then reverted because it broke the
extension's unit tests thanks to the lack of the core change.

Change-Id: I0ad334168bc251f26d2f29b7ea9107b78d019e7a
Depends-On: I778bab2b87971875aed4c804341ad59c1c6e021f