Revert "API: Split description messages into summary + additional text"
Revert as it was manually merged, which by-passed the Depends
requirement. When adding restrictions, we usually merge extension
changes first. But when changing restrictions in a back-compat way
(the current becomes deprecated), the core change must merge
first as otherwise unit tests will be failing until that happens.

Right now, this repo is broken since it is failing unit tests.

This reverts commit a88d0773941e80e66f1f87c77972f4fa2daf6bf2.

Change-Id: I85eee95d5ea4cc31d6c5e1a8262f19f1bc15f5cb