Add README.md file to the repository
A simple read me file with what the extension is all about and
how to go about trying it out.

TODO: Update the read with name "Examples" when the extension
has been renamed from "examples" to "Examples". And REAME.md will
be improved as the extension is updated as well.

Bug: T156568
Change-Id: I89ff1de40eae1e181fb791c356f0a32076f61c42