Do not expect all hook parameters as references
The fact that some code expects the *parser* as a reference is quite
insane, if you start thinking about it. Why would anybody want to
*replace* the parser with another one midway? o_O?

Good news is, this is neither intentional nor an actual risk. It is
effectively a forgotten fragment from a long time ago, when PHP 4
passed objects by value and created possibly expensive clones. It was
an accepted best practice back then to add the & to function parameters
to avoid expensive cloning.

It is pointless nowadays. Luckily it doesn't do much. The code in core
will not even do anything with the changed value when a bogus hook
handler would try to actually replace the parser.

Removing the & from the code that *calls* the hooks is not easily
possible, because this is a breaking change. But it's possible to
remove it from the code *called*.

Let's demonstrate the current best practice and not expect hook
parameters as references when no hook handler is supposed to replace
these objects.

Change-Id: Ic4c33540e614b120d9fa41b96c513f4a9114a18c