Convert ResourceLoaderTestModules hook to QUnitTestModule attribute
This PHP-hook was deprecated in 1.33 in favour of a registration
method that is native to extension.json.
See <https://www.mediawiki.org/wiki/Manual:Hooks/ResourceLoaderTestModules>

Also remove the node-qunit experiment from 0e552715e5, there are
lots of unresolved problems with this, and it is far from being
a recommended practice to be copied by new extensions.

Change-Id: Ib7deb611ba8fb8abd02a884ea5345235f9805e15